% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_for_fmc}
\alias{prep_for_fmc}
\title{Prepare for use with Focal Mechanism Classification}
\usage{
prep_for_fmc(type = NULL)
}
\arguments{
\item{type}{character; the final format}
}
\description{
Prepare for use with Focal Mechanism Classification
}
\references{
url
}

