% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetters.R
\name{salton_trough_yhs}
\alias{salton_trough_yhs}
\title{Extract quakes in the Salton trough}
\usage{
salton_trough_yhs(...)
}
\arguments{
\item{...}{additional parameters to}
}
\description{
Extract quakes in the Salton trough
}
\examples{
salt <- salton_trough_yhs()
plot(Lat.deg ~ Lon.deg, salt, pch=".", asp=1.19)
lines(saltonsea)
}

