% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhs.catalog.R
\docType{data}
\name{saltonsea}
\alias{saltonsea}
\title{Salton Sea polygon}
\format{
\code{'data.frame'}: latitudes and longitudes outlining the shoreline
}
\source{
Digitized from Google-maps in Feb-2015 with \url{http://www.birdtheme.org/useful/v3largemap.html}
}
\description{
Coordinates of a recent Salton Sea shoreline
}
\details{
The first and last coordinates are equal, which means this can be treated as a polygon.
}
\examples{
plot(saltonsea, type='l', asp=1.19)
points(Lat.deg ~ Lon.deg, yhs, pch=".")

# first and last points are equal -- this is a polygon
try(identical(c(saltonsea[1,]), c(saltonsea[nrow(saltonsea),])))
}
\references{
For more information on the Salton Sea:

\url{http://saltonsea.ca.gov/}

\url{http://en.wikipedia.org/wiki/Salton_Trough}
}
\seealso{
\code{\link{yhs.catalog}}
}
