% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhs.catalog.R
\docType{data}
\name{scsn}
\alias{scsn}
\title{SCSN relocated catalog}
\format{
similar to \code{\link{yhs}} but with other details on
the relocation details instead of focal planes
}
\description{
relocated catalog published by the SCSN
}
\details{
this is different from the \code{\link{yhs}} catalog
in that there are no details of focal planes; hence, this catalog
will have many more events that the aforementioned catalog.
}
\references{
\url{http://scedc.caltech.edu/research-tools/alt-2011-dd-hauksson-yang-shearer.html}
}
