% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhs.catalog.R
\docType{package}
\name{yhs.catalog}
\alias{yhs.catalog}
\alias{ysh.catalog}
\alias{yhs.catalog-package}
\alias{ysh.catalog-package}
\title{Refined earthquake focal mechanism catalog for southern California}
\description{
Provides access to the Yang, Hauksson, and Shearer (2012) catalog
of focal mechanisms for southern California.
}
\examples{
data(yhs)
str(yhs)

# plot all locations in space
plot(Lat.deg ~ Lon.deg, yhs, pch=".")
}
\references{
Yang, W., E. Hauksson and P. M. Shearer, 
Computing a large refined catalog of focal mechanisms for southern California 
(1981 - 2010): Temporal Stability of the Style of Faulting, 
Bull. Seismol. Soc. Am., 102(3), 2012.
\url{http://dx.doi.org/10.1785/0120110311}

Hauksson E., W. Yang and P.M. Shearer, 
Waveform relocated earthquake catalog for southern California (1981 - 2011), 
Bull. Seismol. Soc. Am., 102(5), 2012.
\url{http://dx.doi.org/10.1785/0120120010}
}
\seealso{
\code{\link{yhs}}
}
\author{
A.J. Barbour
}
