% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhs.catalog.R
\docType{data}
\name{yhs}
\alias{yhs}
\alias{ysh}
\title{Yang, Hauksson, and Shearer refined focal mechanism catalog for southern California}
\format{
Classes \code{'tbl_df'}, \code{'tbl'} and \code{'data.frame'}:  
196993 obs. of  22 variables in the HASH format: 
179255 earthquakes from the original catalog, and 17738 from the supplemental catalogs.
}
\source{
SCEC: \url{http://scedc.caltech.edu/research-tools/alt-2011-yang-hauksson-shearer.html}
}
\description{
This is the dataset containing the earthquake catalog.
}
\details{
This dataset consists of the original catalog and a supplement. These can be
identified with the \code{src} field: \code{src=='orig'} for the original
and \code{src=='supp'} for the supplemental.
}
\examples{
require(dplyr)
print(yhs)

# source and quality distribution
print(tbl <- with(yhs, table(Source, Quality)))
plot(tbl)

print(tbl2 <- with(yhs, table(Year, Quality)))
plot(tbl2)

# plot all locations in space
plot(Lat.deg ~ Lon.deg, yhs, pch=".")

# highlight the supplemental catalog in red
plot(Lat.deg ~ Lon.deg, yhs, pch=".", col=as.numeric(factor(Source)))
}
\references{
HASH:

(1) Hardebeck, Jeanne L. and Peter M. Shearer, 
A new method for determining first-motion focal mechanisms, 
Bull. Seismol. Soc. Am., 92(6), 2002.
\url{https://doi.org/10.1785/0120010200}

(2) Hardebeck, Jeanne L. and Peter M. Shearer, 
Using S/P Amplitude Ratios to Constrain the Focal Mechanisms of Small Earthquakes, 
Bull. Seismol. Soc. Am., 93(6), 2003.
\url{https://doi.org/10.1785/0120020236}
}
\seealso{
\code{\link{yhs.catalog}}
}
